requirements.txt
dnspython>=2.3.0
requests>=2.31.0
tqdm>=4.65.0